+++ docs-ui/src/app/components/container/snippets.ts
@@ -1,0 +1,19 @@
+export const containerUsageSnippet = `import { Container } from "@backstage/ui";
+
+<Container>Hello World!</Container>`;
+
+export const previewSnippet = `<div style={{ maxWidth: '600px', margin: '0 auto' }}>
+  <DecorativeBox />
+</div>`;
+
+export const containerSimpleSnippet = `<Container>
+  <Box>Hello World</Box>
+  <Box>Hello World</Box>
+  <Box>Hello World</Box>
+</Container>`;
+
+export const containerResponsiveSnippet = `<Container paddingY={{ xs: 'sm', md: 'md' }}>
+  <Box>Hello World</Box>
+  <Box>Hello World</Box>
+  <Box>Hello World</Box>
+</Container>`;