+++ packages/frontend-dynamic-feature-loader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/frontend-dynamic-feature-loader
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.13.2
+
 ## 0.1.8-next.0
 
 ### Patch Changes